requirements.txt
Flask==2.3.3
requests==2.31.0
psutil==5.9.6
pystray==0.19.5
Pillow==10.0.1